-refactor: add coreui module
-add coreui module
-move basic activity and fragment from $presentation to $coreui
-fix package names in manifest for $app, $data, $domain, $presentation modules

Now all custom reusable widgets, layouts, colors, fonts, styles etc should be stored in coreui module
