+++ sources/base/coreui/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="my.company.app.presentation" />
+    package="my.company.app.coreui" />
