+++ sources/base/coreui/src/main/res/values/colors.xml
