+++ sources/base/domain/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="my.company.app"/>
+    package="my.company.app.domain"/>
